small code refactoring
