+++ src/content/search/cssStyling/colors.ts
@@ -1,0 +1,41 @@
+export const ColorCount = 9;
+export const Colors = {
+    accent: `60 100% 55%`, //selected highlight
+    0: {
+        highlight: `270 100% 71%`,  //color of the highlighted match on page
+        dark: `268 61% 9%`,         //UI color for matching search panel
+        light: `270 100% 71%`       //UI color for matching search panel
+    }, 1: {
+        highlight: `330 100% 58%`,
+        dark: `333 90% 15%`,
+        light: `330 100% 58%`
+    }, 2: {
+        highlight: `32 72% 50%`,
+        dark: `30 100% 23%`,
+        light: `22 100% 77%`
+    }, 3: {
+        highlight: `228 100% 43%`,
+        dark: `202 100% 9%`,
+        light: `225 100% 64%`
+    }, 4: {
+        highlight: `180 100% 29%`,
+        dark: `180 100% 10%`,
+        light: `180 100% 33%`
+    }, 5: {
+        highlight: `204 100% 71%`,
+        dark: `204 45% 19%`,
+        light: `208 100% 77%`
+    }, 6: {
+        highlight: `0 96% 51%`,
+        dark: `0 100% 15%`,
+        light: `0 95% 71%`
+    }, 7: {
+        highlight: `330 100% 86%`,
+        dark: `330 18% 20%`,
+        light: `330 100% 86%`
+    }, 8: {
+        highlight: `120 100% 57%`,
+        dark: `120 80% 20%`,
+        light: `120 100% 57%`
+    }
+}
