+++ src/content/search/state.ts
@@ -1,4 +1,4 @@
-import { ColorCount } from './cssStyling/cssInjection'
+import { ColorCount } from './cssStyling/colors'
 
 //Holds all user-editable data about a single search
 //Serialized when passed to background script for caching
@@ -8,22 +8,20 @@ export class State {
     public caseSensitive = false;
     public wholeWord = false;
     constructor(public colorIndex = 0, public searchString = "") {
-
     }
 
     static Load(state: State): State {
-        const result = new State();
-        Object.assign(result, state);
-        return result;
+        return Object.assign(new State(), state);
     }
 
     NextColor(): void {
-        this.colorIndex = this.colorIndex == ColorCount - 1 ? 0 : this.colorIndex + 1;
+        this.colorIndex = (this.colorIndex < ColorCount - 1) ? this.colorIndex + 1 : 0;
     }
 
     GetRegex(escape = true): RegExp {
-        let regexString = escape ? this.searchString.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")
-            : this.searchString;
+        let regexString = escape ?
+            this.searchString.replace(/[.*+?^${}()|[\]\\]/g, "\\$&") :
+            this.searchString;
 
         if (this.wholeWord)
             regexString = `\\b${regexString}\\b`;
@@ -36,7 +34,8 @@ export class State {
         return this.searchString == "";
     }
 
-    //Takes array of states, returns a free colorIndex
+    //Takes array of states, returns a free colorIndex 
+    //if none are free, return next color after the latest
     static GetNextColor(states: State[]): number {
         if (!states || states.length == 0)
             return 0;
